installed animejs. settings modal functional.
